requirements.txt
Django==5.1.2
python-dotenv==1.0.1
django-simple-history==3.5.0
waitress==3.0.0
djangorestframework==3.15.2
pyjwt==2.8.0
psycopg2==2.9.9
drf-spectacular==0.27.2
calamus==0.4.2
python-decouple==3.8
crispy-bootstrap5==2024.10
django-crispy-forms==2.3
